requirements.txt
asgiref==3.2.10
attrs==20.2.0
beautifulsoup4==4.9.1
Brotli==1.0.9
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
cubes==1.1
dash==1.16.0
dash-core-components==1.12.0
dash-html-components==1.1.1
dash-renderer==1.8.0
dash-table==4.10.1
Django==3.1.1
django-adminlte-3==0.1.6
emoji==0.6.0
emosent-py==0.1.6
expressions==0.2.3
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
googletrans==3.0.0
grako==3.99.9
h11==0.10.0
h2==3.2.0
hpack==4.0.0
hstspreload==2020.9.9
httpcore==0.10.2
httpx==0.14.3
hyperframe==6.0.0
idna==2.10
itsdangerous==2.0.0a1
Jinja2==3.0.0a1
joblib==0.16.0
jsonschema==3.2.0
MarkupSafe==2.0.0a1
nltk==3.5
nltkdata==0.0.1
npm==0.1.1
numpy==1.19.2
optional-django==0.3.0
pandas==1.1.2
plotly==4.9.0
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
retrying==1.3.3
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.49.0
urllib3==1.25.10
Werkzeug==1.0.1

bonobo~=0.6.4